requirements.txt
streamlit
numpy
matplotlib
torch
torchvision
Pillow
pyngrok